app/requirements.txt
click==6.7
dj-database-url==0.4.2
Django==2.0.1
django-crispy-forms==1.7.0
gunicorn==19.7.1
psycopg2==2.7.3.2
python-dotenv==0.7.1
pytz==2017.3
whitenoise==3.3.1